5. Adding a UI to your App
Adding a Barcode to your app
Typically you would take pictures of bar codes with the device's camera, or maybe process the camera preview. That takes some coding, and in later steps you'll see a sample that does this. To keep things simple, for this lab, you're just going to process an image that is already present in your app.

https://codelabs.developers.google.com/codelabs/bar-codes/#4